Finished exercise 05 of Foundations/flex
Initial solution involved displaying the ! icon, a div containing
header, text and bottom buttons, and the X icon as a flex
container. I set the middle div to flex and set the direction to
column with a gap. Formatted header text, justified to space
around and aligned the items to the flex-start. This got me
nearly there, though I was missing the text bleeding into the X
button area, since it was in its own column flex container. I also
used min-width to prevent corner bottoms from shrinking, where
`flex-shrink: 0` would have made more sense.

After checking solution, realized the modal could be the initial
flex container, everything except the ! icon could be made into a
container (so modal has 2 items), and within that container, the
header div could include the X icon to separate them with space
between while displaying the rest of the information in blocks,
which would achieve the same result without the hard boundary of
the middle column container.