Finished exercise 06 in Foundations/flex
Turned body into flex container, set it as column. This puts
header at the top, content in the middle, footer at the bottom.
Turned both header and footer into flex container, set justify
to space-between, since the left and right side were already
div'd. Eyeballed 16px of padding on both of these. Set footer bg
color to one provided. Reset margin and padding for ul, set
list-style to none to get rid of bullets. Set display to flex to
make horizontal, set a gap of 16px to match padding. Set anchor
text decoration to none to get rid of underline, and color to dark
slate gray to avoid the blue/purple links. Found through solution
that color was #666 instead. Set content to align self at center,
display to flex, direction to column and align items to center
with a 16px gap. Set button padding to 8px 16px.

Solution was almost the same as mine, though instead of justifying
content on the body and aligning the content div to center with
align self, they set flex 1 on content to make it take up the rest
of the space, justified content and aligned items. This achieved
the same result with less redundancy.