Finished exercise 07 of Foundations/flex
Set the body as a flex container column. Made a div around the
sidebar+cards, and another around just the cards. Formatted header
and footer accordingly using flex to vertical center them. Set
sidebar to min-width. Set cards container to flex wrap, with a width
of 300 in the cards so the wrap activates. Set appropriate padding
and gap. The only change from solution was that height was unnecesary
in the cards because flex wrap takes care of it, and the padding on
the header title was set to 0 16px instead of padding-left 16px the
way I had done it. Probably a more stable solution, though practically
identical.